+++ src/lib/components/Profile/ProfileMenu.svelte
@@ -4,7 +4,8 @@
     import { userDetailStore } from '$lib/stores/stores';
     import { get } from 'svelte/store';
     import { reloadStore } from '$lib/stores/reload';
-  
+    import { BACKEND_MEDIA_URL } from '$lib/api/api';
+
     let showDropdown = false;
     let dropdownRef: HTMLElement;
   
@@ -49,7 +50,11 @@
   <div class="relative" bind:this={dropdownRef}>
     <button on:click={toggleDropdown}>
       <!-- TODO: Change avatar -->
-      <iconify-icon class="text-4xl px-4" icon="healthicons:ui-user-profile"></iconify-icon>
+      {#if $userDetailStore && $userDetailStore.avatar}
+        <div class="px-4"><img class="w-[2rem] h-[2rem] object-cover" alt="" src={`${BACKEND_MEDIA_URL}/${$userDetailStore.avatar}`}></div>
+      {:else}
+        <iconify-icon class="text-4xl px-4" icon="healthicons:ui-user-profile"></iconify-icon>
+      {/if }
     </button>
   
     {#if showDropdown}
